Strongly typed all GlobalKeys that reference ProseTextBlock, remove remaining remnants of SuperSelectableText from super_text package
