+++ super_editor/lib/src/infrastructure/super_textfield/android/_editing_controls.dart
@@ -60,9 +60,7 @@ class AndroidEditingOverlayControls extends StatefulWidget {
 
   /// [GlobalKey] that references the widget that contains the text within
   /// the text field.
-  ///
-  /// The referenced widget's `State` object must implement [ProseTextBlock].
-  final GlobalKey textContentKey;
+  final GlobalKey<ProseTextState> textContentKey;
 
   /// The color of the selection handles.
   final Color handleColor;
@@ -152,7 +150,7 @@ class _AndroidEditingOverlayControlsState extends State<AndroidEditingOverlayCon
     });
   }
 
-  ProseTextLayout get _textLayout => (widget.textContentKey.currentState as ProseTextBlock).textLayout;
+  ProseTextLayout get _textLayout => widget.textContentKey.currentState!.textLayout;
 
   void _rebuildOnNextFrame() {
     // We request a rebuild at the end of this frame so that the editing
