+++ super_editor/lib/src/infrastructure/super_textfield/ios/ios_textfield.dart
@@ -140,7 +140,7 @@ class _SuperIOSTextFieldState extends State<SuperIOSTextField> with SingleTicker
   final _textFieldLayerLink = LayerLink();
   final _textContentLayerLink = LayerLink();
   final _scrollKey = GlobalKey<IOSTextFieldTouchInteractorState>();
-  final _textContentKey = GlobalKey();
+  final _textContentKey = GlobalKey<ProseTextState>();
 
   late FocusNode _focusNode;
 
@@ -345,7 +345,7 @@ class _SuperIOSTextFieldState extends State<SuperIOSTextField> with SingleTicker
   }
 
   void _onFloatingCursorChange(RawFloatingCursorPoint point) {
-    _floatingCursorController.updateFloatingCursor((_textContentKey.currentState! as ProseTextBlock).textLayout, point);
+    _floatingCursorController.updateFloatingCursor(_textContentKey.currentState!.textLayout, point);
   }
 
   @override
