+++ super_text/lib/super_text.dart
@@ -1,5 +1,4 @@
 export 'src/caret_layer.dart';
-export 'src/super_selectable_text.dart';
 export 'src/super_text.dart';
 export 'src/super_text_with_selection.dart';
 export 'src/text_layout.dart';
